cephdriver: support snapshots, move volume tooling to own file
